tables-exercise

+++ albums_migration.sql
@@ -5,7 +5,7 @@ CREATE TABLE albums(
     id INT UNSIGNED NOT NULL AUTO_INCREMENT,
     artist_name VARCHAR(100) NOT NULL,
     record_name VARCHAR(100) NOT NULL,
-    release_date SMALLINT UNSIGNED,
+    release_date INT UNSIGNED,
     sales FLOAT NOT NULL,
     genre VARCHAR(100) NOT NULL,
     PRIMARY KEY (id)
